removed extra functions from opticam_analyse to misc also cleaned a bit in opticam_pipe
